Restore specialized detector dispatch (#8)
* fix: restore specialized detector dispatch

* chore(release): finalize 0.3.0 notes

+++ src/pisama_detectors/detection/cost.py
@@ -80,7 +80,7 @@ class CostResult:
     input_cost_usd: float
     output_cost_usd: float
     total_cost_usd: float
-    total_cost_cents: int
+    total_cost_cents: float
     model: str
     provider: str
